backend/pyproject.toml
[tool.poetry]
name = "fax"
version = "0.1.0"
description = ""
authors = ["aslan <baristan1aslan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-rest-framework = "^0.1.0"
django = "^4.1.4"
django-extensions = "^3.2.1"
isort = "^5.11.4"
djoser = "^2.1.0"
drf-yasg = "^1.21.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
